Move http handlers to a separate file
